Move possibleImportsCache to separate file and remove some commented out code
